refactor: rename Static.tsx → Derive.tsx and declareStatic.ts → derive.ts
