+++ packages/core/src/static/derive.ts
