+++ packages/core/src/static/index.ts
@@ -1,6 +1,6 @@
 export * from './decodeVars';
 export * from './declareVar';
-export * from './declareStatic';
+export * from './derive';
 export * from './indexVars';
 export * from './extractVars';
 export * from './condenseVars';
