+++ packages/react-core/src/index.ts
@@ -9,7 +9,7 @@ import Var from './variables/Var';
 import Num from './variables/Num';
 import Currency from './variables/Currency';
 import DateTime from './variables/DateTime';
-import { Static, Derive } from './variables/Static';
+import { Static, Derive } from './variables/Derive';
 import Plural from './branches/plurals/Plural';
 import Branch from './branches/Branch';
 import useLocales from './hooks/useLocales';
